refactor: useInputRefs 훅을 추가하여 코드 중복 제거
